requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector-python==8.2.0
protobuf==4.21.12
Pygments==2.17.2
python-dotenv==1.0.0
rich==13.7.0
simple-term-menu==1.6.4
